Add checksum to serialization format
The last byte of the serialized format is a checksum, so if any of the
bytes in the message change deserialization will fail. The checksum
algorithm is taken from the function used in cansw_radio. This
algorithm is position independant, so this won't detect byte order
swaps, but will detect if any bits are changed in the datastream. This
isn't great, but it's better than nothing, and works for v1.0.0.